comment out error-inducing code in mam4_amicphys.cpp and small reorg change in online_emission.hpp
