+++ components/eamxx/src/physics/mam/impl/mam4_amicphys.cpp
@@ -677,13 +677,13 @@ void mam_amicphys_1subarea_clear(
       const int ntot_soamode = 4;
       int niter_out = 0;
       Real g0_soa_out = 0;
-      gasaerexch::mam_gasaerexch_1subarea(
-          nghq, igas_h2so4, igas_nh3, ntot_soamode, gas_to_aer, iaer_so4,
-          iaer_pom, l_calc_gas_uptake_coeff, l_gas_condense_to_mode,
-          eqn_and_numerics_category, dtsubstep, dtsub_soa_fixed, temp, pmid,
-          aircon, num_gas_ids, qgas_cur, qgas_avg, qgas_netprod_otrproc,
-          qaer_cur, qnum_cur, dgn_awet, alnsg_aer, uptk_rate_factor, uptkaer,
-          uptkrate_h2so4, niter_out, g0_soa_out);
+      // gasaerexch::mam_gasaerexch_1subarea(
+      //     nghq, igas_h2so4, igas_nh3, ntot_soamode, gas_to_aer, iaer_so4,
+      //     iaer_pom, l_calc_gas_uptake_coeff, l_gas_condense_to_mode,
+      //     eqn_and_numerics_category, dtsubstep, dtsub_soa_fixed, temp, pmid,
+      //     aircon, num_gas_ids, qgas_cur, qgas_avg, qgas_netprod_otrproc,
+      //     qaer_cur, qnum_cur, dgn_awet, alnsg_aer, uptk_rate_factor, uptkaer,
+      //     uptkrate_h2so4, niter_out, g0_soa_out);
 
       if (config.newnuc_h2so4_conc_optaa == 11)
         qgas_avg[igas_h2so4] =
@@ -823,17 +823,67 @@ void mam_amicphys_1subarea_clear(
       Real dnh4dt_ait = 0;
       Nucleation nucleation;
       Nucleation::Config config;
-      config.dens_so4a_host = 1770;
+      // config.dens_so4a_host = 1770;
       config.mw_nh4a_host = 115;
       config.mw_so4a_host = 115;
-      config.accom_coef_h2so4 = 0.65;
+      // config.accom_coef_h2so4 = 0.65;
       AeroConfig aero_config;
       nucleation.init(aero_config, config);
-      nucleation.compute_tendencies_(
-          dtsubstep, temp, pmid, aircon, zmid, pblh, relhum, uptkrate_h2so4,
-          del_h2so4_gasprod, del_h2so4_aeruptk, qgas_cur, qgas_avg, qnum_cur,
-          qaer_cur_tmp, qwtr_cur, dndt_ait, dmdt_ait, dso4dt_ait, dnh4dt_ait,
-          dnclusterdt_substep);
+      // new version after switching from box model version of nucleation
+      // compute_tendencies_(
+      //                     // Real deltat,
+      //                     // Real temp_in,
+      //                     // Real press_in,
+      //                     // Real zm_in,
+      //                     // Real pblh_in,
+      //                     // Real relhum,
+      //                     // Real uptkrate_h2so4,
+      //                     const int nsize, // missing?
+      //                     const Real dp_lo_mode,
+      //                     const Real dp_hi_mode,
+      //                     // const Real qgas_cur[num_gases],
+      //                     // const Real qgas_avg[num_gases],
+      //                     int &isize_nuc,
+      //                     // Real &qnuma_del,
+      //                     // Real &qso4a_del,
+      //                     // Real &qnh4a_del,
+      //                     // Real &qh2so4_del,
+      //                     Real &qnh3_del,
+      //                     // Real &dnclusterdt
+      //                     )
+      // old, box-model-based version
+      // nucleation.compute_tendencies_(
+      //     dtsubstep, temp, pmid, aircon, zmid, pblh, relhum, uptkrate_h2so4,
+      //     del_h2so4_gasprod, del_h2so4_aeruptk, qgas_cur, qgas_avg, qnum_cur,
+      //     qaer_cur_tmp, qwtr_cur, dndt_ait, dmdt_ait, dso4dt_ait, dnh4dt_ait,
+      //     dnclusterdt_substep);
+      // nucleation.compute_tendencies_(
+      //     dtsubstep, // deltat
+      //     temp, // temp_in
+      //     pmid, // press_in
+      //     aircon, // ?
+      //     zmid, // zm_in
+      //     pblh, // pblh_in
+      //     relhum, // relhum
+      //     uptkrate_h2so4, // uptkrate_h2so4
+      //     // nsize?
+      //     // dp_lo_mode?
+      //     // dp_hi_mode?
+      //     del_h2so4_aeruptk, // uptkrate_h2so4?
+      //     qgas_cur, // qgas_cur[num_gases]?
+      //     qgas_avg, // qgas_avg[num_gases]?
+      //     // isize_nuc?
+      //     qnum_cur, // qnuma_del?
+      //     qaer_cur_tmp, // ?
+      //     qwtr_cur, // ?
+      //     dndt_ait, //
+      //     dmdt_ait, //
+      //     dso4dt_ait, // qso4a_del
+      //     dnh4dt_ait, // qnh4a_del
+      //     del_h2so4_gasprod, // qh2so4_del?
+      //     // qnh3_del?
+      //     dnclusterdt_substep // dnclusterdt
+      //     );
       for (int j = 0; j < num_aerosol_ids; ++j)
         for (int i = 0; i < num_modes; ++i)
           qaer_cur[j][i] = qaer_cur_tmp[i][j];
@@ -879,9 +929,9 @@ void mam_amicphys_1subarea_clear(
       for (int j = 0; j < num_aerosol_ids; ++j)
         for (int i = 0; i < num_modes; ++i)
           qaer_sv1[j][i] = qaer_cur[j][i];
-      coagulation::mam_coag_1subarea(dtsubstep, temp, pmid, aircon, dgn_a,
-                                     dgn_awet, wetdens, qnum_cur, qaer_cur,
-                                     qaer_delsub_coag_in);
+      // coagulation::mam_coag_1subarea(dtsubstep, temp, pmid, aircon, dgn_a,
+      //                                dgn_awet, wetdens, qnum_cur, qaer_cur,
+      //                                qaer_delsub_coag_in);
       for (int i = 0; i < num_modes; ++i)
         qnum_delsub_coag[i] = qnum_cur[i] - qnum_sv1[i];
       for (int j = 0; j < num_aerosol_ids; ++j)
@@ -1146,13 +1196,13 @@ void mam_amicphys_1subarea_cloudy(
       Real g0_soa_out = 0;
       // time sub-step
       const Real dtsub_soa_fixed = -1.0;
-      gasaerexch::mam_gasaerexch_1subarea(
-          nghq, igas_h2so4, igas_nh3, ntot_soamode, gas_to_aer, iaer_so4,
-          iaer_pom, l_calc_gas_uptake_coeff, l_gas_condense_to_mode,
-          eqn_and_numerics_category, dtsubstep, dtsub_soa_fixed, temp, pmid,
-          aircon, num_gas_ids, qgas_cur, qgas_avg, qgas_netprod_otrproc,
-          qaer_cur, qnum_cur, dgn_awet, alnsg_aer, uptk_rate_factor, uptkaer,
-          uptkrate_h2so4, niter_out, g0_soa_out);
+      // gasaerexch::mam_gasaerexch_1subarea(
+      //     nghq, igas_h2so4, igas_nh3, ntot_soamode, gas_to_aer, iaer_so4,
+      //     iaer_pom, l_calc_gas_uptake_coeff, l_gas_condense_to_mode,
+      //     eqn_and_numerics_category, dtsubstep, dtsub_soa_fixed, temp, pmid,
+      //     aircon, num_gas_ids, qgas_cur, qgas_avg, qgas_netprod_otrproc,
+      //     qaer_cur, qnum_cur, dgn_awet, alnsg_aer, uptk_rate_factor, uptkaer,
+      //     uptkrate_h2so4, niter_out, g0_soa_out);
 
       if (config.newnuc_h2so4_conc_optaa == 11)
         qgas_avg[igas_h2so4] =
